refactor(cli): split service plist rendering into sibling module
`crates/convergio-cli/src/commands/service.rs` was at 288 lines
inside the 300-line cap but with too little headroom to add the
regression guard for the upcoming `RunAtLoad` / `KeepAlive`
defaults fix (audit follow-up T828d03c). Move the launchd /
systemd unit rendering and their existing tests into a new
`service_unit` sibling module so the orchestration file stays
under the cap.

No behavior  `current()` still emits the same plist /change
unit content. Tests previously living in `service::tests` now
live in `service_unit::tests`.

Tracks: T828d03c7-4019-4c60-9a13-88f2404e1f9d
