+++ crates/convergio-cli/src/commands/mod.rs
@@ -51,6 +51,7 @@ mod plan_run;
 mod plan_triage;
 pub mod pr;
 pub mod service;
+pub(crate) mod service_unit;
 pub mod session;
 pub mod setup;
 mod setup_agent_prompt;
